+++ server/src/main/kotlin/com/handcontrol/server/mqtt/command/dto/TelemetryDto.kt
@@ -10,18 +10,18 @@ import kotlinx.serialization.Serializable
 @Serializable
 @ExperimentalUnsignedTypes
 data class TelemetryDto(
-        val telemetryFrequency: Short = 0,
+        val telemetryFrequency: Int = 0,
         val emgStatus: ModuleStatusType = ModuleStatusType.MODULE_STATUS_INITIALIZATION,
         val displayStatus: ModuleStatusType = ModuleStatusType.MODULE_STATUS_INITIALIZATION,
         val gyroStatus: ModuleStatusType = ModuleStatusType.MODULE_STATUS_INITIALIZATION,
         val driverStatus: DriverStatusType = DriverStatusType.DRIVER_STATUS_INITIALIZATION,
         val lastTimeSync: Long = 0,
-        val emg: Short = 0,
+        val emg: Int = 0,
         val executableGesture: UuidDto = UuidDto(""),
-        val power: Byte = 0,
-        val pointerFingerPosition: Byte = 0,
-        val middleFingerPosition: Byte = 0,
-        val ringFinderPosition: Byte = 0,
-        val littleFingerPosition: Byte = 0,
-        val thumbFingerPosition: Byte = 0
+        val power: Int = 0,
+        val pointerFingerPosition: Int = 0,
+        val middleFingerPosition: Int = 0,
+        val ringFinderPosition: Int = 0,
+        val littleFingerPosition: Int = 0,
+        val thumbFingerPosition: Int = 0
 )
